docker-compose: only publish and expose required ports
Some ports are only needed inside the network shared between the
front-end and the back-end containers.

This is a partial regression from #5c94384a.

Using localhost to make requests on the development machine, where the
browser and the docker containers were run, led to the false asumption
that exposing those ports would make things work. And they did work,
until the app was tested on the target, which is not the same as the web
browser making the requests, and thus localhost sent the requests to an
unexistent URL.